Moved special tags source definitions from movepiece.ts -> moveutil.ts
